Flatten contentScript
