+++ src/contentScript/lyraApi/getPublicKey.ts
