+++ src/contentScript/lyraApi/requestSignature.ts
