+++ src/helpers/stellar.ts
@@ -1,0 +1,2 @@
+export const truncatedPublicKey = (publicKey: string) =>
+  `${publicKey.slice(0, 4)}...${publicKey.slice(-4)}`;
